requirements.txt
Pillow>=10.0.0
torch>=2.2.0
transformers>=4.40.0
PyYAML>=6.0.0
pytest>=8.0.0
ruff>=0.6.0
